Refactor WebRTC functions into a separate file
Moved all WebRTC-related functions and variables from `js/main.js` to a new file, `js/web-rtc.js`. Updated `index.html` to include the new script. This change improves code organization and modularity without altering functionality.
