+++ index.html
@@ -266,6 +266,7 @@
     <script src="js/video-player.js"></script>
     <script src="js/worker.js"></script>
     <script src="js/mobs.js"></script>
+    <script src="js/web-rtc.js"></script>
     <script src="js/main.js"></script>
 </body>
 </html>
